plugins/requirements.txt
numpy
fastapi
jsonschema
PyYAML
pre-commit
